+++ python/lsst/ts/m2gui/application.py
@@ -32,23 +32,32 @@ from PySide2.QtCore import QCommandLineOption, QCommandLineParser
 from .main_window import MainWindow
 
 
-def run_application(argv: list) -> None:
-    """Run the application.
-
-    Parameters
-    ----------
-    argv : `list`
-        Arguments from the command line.
-    """
+def run_application() -> None:
+    """Run the application."""
 
     if "QT_API" not in os.environ:
         os.environ.setdefault("QT_API", "PySide2")
         print("qasync: QT_API not set, defaulting to PySide2.")
 
-    try:
-        qasync.run(main(argv))
-    except asyncio.exceptions.CancelledError:
-        sys.exit(0)
+    # Workaround the Python 3.11 issue in 'qasync' module based on:
+    # https://github.com/CabbageDevelopment/qasync/issues/68
+    if (sys.version_info.major == 3) and (sys.version_info.minor >= 11):
+        with qasync._set_event_loop_policy(qasync.DefaultQEventLoopPolicy()):
+            runner = asyncio.runners.Runner()
+
+            try:
+                runner.run(main(sys.argv))
+
+            except asyncio.exceptions.CancelledError:
+                sys.exit(0)
+
+            finally:
+                runner.close()
+    else:
+        try:
+            qasync.run(main(sys.argv))
+        except asyncio.exceptions.CancelledError:
+            sys.exit(0)
 
 
 async def main(argv: list) -> bool: