+++ doc/version_history.rst
@@ -6,6 +6,18 @@
 Version History
 ##################
 
+.. _lsst.ts.m2gui-0.6.3:
+
+-------------
+0.6.3
+-------------
+
+* Add the **Jenkinsfile.conda** and **setup.py**.
+* Move the policy files to **ts_config_mttcs**.
+* Remove the ``log/`` directory and output the log file to ``/rubin/mtm2/log`` if possible.
+* Remove the argument of ``run_application()``.
+* Workaround the Python 3.11 issue in **qasync** module.
+
 .. _lsst.ts.m2gui-0.6.2:
 
 -------------